Updated the submit.py file to now work with Molpro on Zeus
Changed bin/sh to bin/bash

Added initial/final time

Add the flag  to the bin/bash commands in case the user requires their profile to be loaded as while. Addtionally, added an extra space at the end of the script

Fixed the cp commands that did not have Target Dest. Also added if/else statement for when geometry files exist

Added Hash

Forgot to add hash in front of bin/bash